Trivial Changes - Added a default to the APT hook switch statement - Fixed various incorrect print format specifiers - Fixed various macro redefinitions - Fixed some missing return statements - Fixed some uninitialized variables - Fixed a strange warning for the SSettings variable's initialization - Fixed some missing parentheses and brackets - Fixed some incorrect struct typedefs - Removed some unknown pragmas from fast_gaussian_blur_template - Fixed a name conflict between the _C macro in cheats.cpp and one in STL caused by _C being a reserved name. - Removed some function prototypes with no implementations

+++ source/3dsmain.cpp
@@ -1287,7 +1287,7 @@ bool settingsReadWriteFullListByGame(bool writeMode)
     }
 
     char version[10];
-    snprintf(version, sizeof(version), "%.1f", GAME_CONFIG_FILE_TARGET_VERSION);
+    snprintf(version, sizeof(version), "%.1f", (double) GAME_CONFIG_FILE_TARGET_VERSION);
     config3dsReadWriteString(stream, writeMode, "#v%s\n", "#v%10[^\n]\n", version);
     float detectedConfigVersion = config3dsGetVersionFromFile(writeMode, true, version);
